feat: :children_crossing: Better handling of error situations in Settings view
Added not-so-brutal checks into Settings view for from and to defaults.
Now user gets nice error message when from and/or to value is missing.
It is also checked that from value can't be greater than to value.
