+++ src/components/worklogInputs.tsx
@@ -3,6 +3,7 @@ import { Dispatch, SetStateAction } from 'react';
 import TimeInput from './form/timeInput';
 import Checkbox from './form/checkbox';
 import { Date_Time } from '@/util/dateFormatter';
+import { timeIsGt } from '@/util/date';
 
 export default function WorklogInputs({
   value,
@@ -11,30 +12,16 @@ export default function WorklogInputs({
   value: WorklogFormDataEntry;
   setValue: Dispatch<SetStateAction<WorklogFormDataEntry>>;
 }) {
-  const isGt = (a: Date_Time, b: Date_Time) => {
-    const a_split = a.split(':');
-    const b_split = b.split(':');
-
-    if (parseInt(a_split[0]) > parseInt(b_split[0])) {
-      return true;
-    } else if (parseInt(a_split[0]) === parseInt(b_split[0])) {
-      if (parseInt(a_split[1]) > parseInt(b_split[1])) {
-        return true;
-      }
-    }
-    return false;
-  };
-
   const updateFrom = (time?: Date_Time) => {
     const newFrom = time ?? '';
     const newTo =
-      newFrom && value.to && isGt(newFrom, value.to) ? newFrom : value.to;
+      newFrom && value.to && timeIsGt(newFrom, value.to) ? newFrom : value.to;
     setValue({ ...value, from: newFrom, to: newTo });
   };
   const updateTo = (time?: Date_Time) => {
     const newTo = time ?? '';
     const newFrom =
-      newTo && value.from && isGt(value.from, newTo) ? newTo : value.from;
+      newTo && value.from && timeIsGt(value.from, newTo) ? newTo : value.from;
     setValue({ ...value, from: newFrom, to: newTo });
   };
